please be QUIET!

+++ src/apps/chem/SCF.cc
@@ -2226,7 +2226,7 @@ void SCF::update_subspace(World & world, vecfuncT & Vpsia, vecfuncT & Vpsib,
 				END_TIMER(world, "Subspace transform");
 				if (param.maxsub() <= 1) {
 					subspace.clear();
-				} else if (subspace.size() == param.maxsub()) {
+				} else if (subspace.size() == size_t(param.maxsub())) {
 					subspace.erase(subspace.begin());
 					Q = Q(Slice(1, -1), Slice(1, -1));
 				}
@@ -2958,7 +2958,7 @@ void SCF::update_response_subspace(World & world,
 	END_TIMER(world, "Subspace transform");
 	if(param.maxsub() <= 1){
 		subspace.clear();
-	} else if(subspace.size() == param.maxsub()){
+	} else if(subspace.size() == size_t(param.maxsub())){
 		subspace.erase(subspace.begin());
 		Q = Q(Slice(1, -1), Slice(1, -1));
 	}
